+++ src/transport/mqtt/connection.rs
@@ -50,11 +50,11 @@ use std::{
 };
 
 use rumqttc::{
-    mqttbytes, ClientError, ConnectionError, Event, NoticeError, Packet, Publish, QoS, StateError,
+    mqttbytes, ClientError, ConnectionError, Event, Packet, Publish, QoS, StateError, TokenError,
     Transport,
 };
 use sync_wrapper::SyncWrapper;
-use tokio::task::JoinHandle;
+use tokio::task::{JoinError, JoinHandle};
 use tracing::{debug, error, info, trace, warn};
 
 use crate::{
@@ -69,7 +69,7 @@ use crate::{
 use super::{
     client::{AsyncClient, EventLoop},
     config::transport::TransportProvider,
-    ClientId, PayloadError, SessionData,
+    ClientId, PairingError, PayloadError, SessionData,
 };
 
 /// Errors while initializing the MQTT connection.
@@ -85,9 +85,9 @@ enum InitError {
     },
     /// Couldn't wait for the message acknowledgment.
     #[error("couldn't wait the message acknowledgment for {ctx}")]
-    Notice {
+    Ack {
         #[source]
-        backtrace: NoticeError,
+        backtrace: TokenError,
         ctx: &'static str,
     },
     /// Couldn't serialize the device property payload.
@@ -101,13 +101,25 @@ enum InitError {
     PurgeProperties(#[source] PropertiesError),
 }
 
+/// Error while polling the connection
+#[non_exhaustive]
+#[derive(thiserror::Error, Debug)]
+pub enum PollError {
+    /// Couldn't reconnect to Astarte
+    #[error("couldn't reconnect to Astarte")]
+    Pairing(#[from] PairingError),
+    /// Couldn't complete a store operation
+    #[error("store operation failed")]
+    Store(#[from] StoreError),
+}
+
 impl InitError {
     const fn client(ctx: &'static str) -> impl Fn(ClientError) -> InitError {
         move |backtrace: ClientError| InitError::Client { backtrace, ctx }
     }
 
-    const fn notice(ctx: &'static str) -> impl Fn(NoticeError) -> InitError {
-        move |backtrace: NoticeError| InitError::Notice { backtrace, ctx }
+    const fn ack(ctx: &'static str) -> impl Fn(TokenError) -> InitError {
+        move |backtrace: TokenError| InitError::Ack { backtrace, ctx }
     }
 }
 
@@ -181,7 +193,7 @@ impl MqttConnection {
         client_id: ClientId<&str>,
         interfaces: &Interfaces,
         store: &StoreWrapper<S>,
-    ) -> Result<Self, StoreError>
+    ) -> Result<Self, PollError>
     where
         S: PropertyStore,
     {
@@ -200,7 +212,7 @@ impl MqttConnection {
         client_id: ClientId<&str>,
         interfaces: &Interfaces,
         store: &StoreWrapper<S>,
-    ) -> Result<(), StoreError>
+    ) -> Result<(), PollError>
     where
         S: PropertyStore,
     {
@@ -215,6 +227,8 @@ impl MqttConnection {
                 .await?;
 
             if let Some(publish) = opt_publish {
+                debug!("received a publish");
+
                 disable_clean_session(self.connection.eventloop_mut());
 
                 self.buff.push_back(publish);
@@ -292,14 +306,14 @@ impl State {
         client_id: ClientId<&str>,
         interfaces: &Interfaces,
         store: &StoreWrapper<S>,
-    ) -> Result<Option<Publish>, StoreError>
+    ) -> Result<Option<Publish>, PollError>
     where
         S: PropertyStore,
     {
         trace!("state {}", self);
 
         let next = match self {
-            State::Disconnected(disconnected) => disconnected.reconnect(conn, client_id).await,
+            State::Disconnected(disconnected) => disconnected.reconnect(conn, client_id).await?,
             State::Connecting(connecting) => connecting.wait_connack(conn).await,
             State::Handshake(handshake) => {
                 let session_data = SessionData::try_from_props(interfaces, store).await?;
@@ -363,21 +377,19 @@ impl Disconnected {
     /// It recreates the credentials and reconnect to the broker, using the same
     /// session. If it fails, it returns an error so that the whole connection process can
     /// be retried.
-    async fn reconnect(&mut self, conn: &mut Connection, client_id: ClientId<&str>) -> Next {
-        let api = ApiClient::new(
-            client_id.realm,
-            client_id.device_id,
-            conn.provider.pairing_url().clone(),
-            conn.provider.credential_secret().to_string(),
-            conn.provider.api_tls_config(),
-        );
+    async fn reconnect(
+        &mut self,
+        conn: &mut Connection,
+        client_id: ClientId<&str>,
+    ) -> Result<Next, PairingError> {
+        let api = ApiClient::from_transport(&conn.provider, client_id.realm, client_id.device_id)?;
 
         let transport = match conn.provider.recreate_transport(&api).await {
             Ok(transport) => transport,
             Err(err) => {
                 error!(error = %Report::new(err),"couldn't pair device");
 
-                return Next::Same;
+                return Ok(Next::Same);
             }
         };
 
@@ -387,7 +399,7 @@ impl Disconnected {
         eventloop.clean();
         Self::set_transport(eventloop, transport);
 
-        Next::state(Connecting)
+        Ok(Next::state(Connecting))
     }
 
     #[cfg(not(test))]
@@ -482,9 +494,8 @@ impl Handshake {
                 .send_introspection(client_id, session_data.interfaces)
                 .await
                 .map_err(InitError::client("send introspection"))?
-                .wait_async()
                 .await
-                .map_err(InitError::notice("subscribe server interface"))?;
+                .map_err(InitError::ack("subscribe server interface"))?;
 
             debug!("session present {}", self.session_present);
             if !self.session_present {
@@ -523,9 +534,8 @@ impl Handshake {
             )
             .await
             .map_err(InitError::client("subscribe consumer properties"))?
-            .wait_async()
             .await
-            .map_err(InitError::notice("subscribe consumer properties"))?;
+            .map_err(InitError::ack("subscribe consumer properties"))?;
 
         debug!(
             "subscribing on {} server interfaces",
@@ -539,9 +549,8 @@ impl Handshake {
 
         if let Some(notice) = notice {
             notice
-                .wait_async()
                 .await
-                .map_err(InitError::notice("subscribe server interface"))?;
+                .map_err(InitError::ack("subscribe server interface"))?;
         }
 
         Ok(())
@@ -563,9 +572,10 @@ impl Handshake {
             )
             .await
             .map_err(InitError::client("empty cache"))?
-            .wait_async()
             .await
-            .map_err(InitError::notice("empty cache"))
+            .map_err(InitError::ack("empty cache"))?;
+
+        Ok(())
     }
 
     /// Sends the passed device owned properties
@@ -590,9 +600,10 @@ impl Handshake {
             )
             .await
             .map_err(InitError::client("purge device properties"))?
-            .wait_async()
             .await
-            .map_err(InitError::notice("purge device properties"))
+            .map_err(InitError::ack("purge device properties"))?;
+
+        Ok(())
     }
 
     /// Sends the passed device owned properties
@@ -624,9 +635,8 @@ impl Handshake {
                 .publish(topic, rumqttc::QoS::ExactlyOnce, false, payload)
                 .await
                 .map_err(InitError::client("device property"))?
-                .wait_async()
                 .await
-                .map_err(InitError::notice("device properties"))?;
+                .map_err(InitError::ack("device properties"))?;
 
             if prop.value.is_none() {
                 trace!("clearing unset property {}/{}", prop.interface, prop.path);
@@ -661,31 +671,30 @@ impl WaitAcks {
     /// packets in are sent correctly and the handle can advance to completion. Eventual published
     /// packets are returned.
     async fn wait_connection(&mut self, conn: &mut Connection) -> Result<Next, StoreError> {
-        // HACK: This is to make the function pass with mockall, since the poll function would never
-        //       yield while using the mocks. With the real implementation the `Eventloop::poll`
-        //       would have to wait for the connection and the sent packets.
-        #[cfg(test)]
-        tokio::task::yield_now().await;
+        tokio::select! {
+            // Join handle is cancel safe
+            res = &mut self.handle => {
+                debug!("task joined");
 
-        if self.handle.is_finished() {
-            debug!("init task finished");
+                Self::handle_join(res)
+            }
+            // I hope this is cancel safe
+            res = conn.eventloop_mut().poll() => {
+                debug!("next event polled");
 
-            return self.handle_join().await;
+                let next = match res {
+                    Ok(event) => Next::handle_event(event),
+                    Err(err) => Next::handle_error(err),
+                };
+
+                Ok(next)
+            }
         }
-
-        trace!("polling next event");
-
-        let next = match conn.eventloop_mut().poll().await {
-            Ok(event) => Next::handle_event(event),
-            Err(err) => Next::handle_error(err),
-        };
-
-        Ok(next)
     }
 
-    async fn handle_join(&mut self) -> Result<Next, StoreError> {
+    fn handle_join(res: Result<Result<(), InitError>, JoinError>) -> Result<Next, StoreError> {
         // Don't move the handle to await the task
-        match (&mut self.handle).await {
+        match res {
             Ok(Ok(())) => {
                 info!("device connected");
                 Ok(Next::state(Connected))
@@ -852,13 +861,15 @@ impl Next {
 mod tests {
     use std::{str::FromStr, time::Duration};
 
+    use mockall::predicate;
+    use rumqttc::{AckOfPub, SubAck};
+
     use crate::{
         store::{memory::MemoryStore, StoredProp},
         test::{DEVICE_PROPERTIES, INDIVIDUAL_SERVER_DATASTREAM, OBJECT_DEVICE_DATASTREAM},
         transport::mqtt::test::notify_success,
         AstarteType, Interface,
     };
-    use mockall::predicate;
 
     use super::*;
 
@@ -875,12 +886,16 @@ mod tests {
             .once()
             .in_sequence(&mut seq)
             .returning(|| {
-                Ok(Event::Incoming(rumqttc::Packet::ConnAck(
-                    rumqttc::ConnAck {
-                        session_present: false,
-                        code: rumqttc::ConnectReturnCode::Success,
-                    },
-                )))
+                Box::pin(async {
+                    tokio::task::yield_now().await;
+
+                    Ok(Event::Incoming(rumqttc::Packet::ConnAck(
+                        rumqttc::ConnAck {
+                            session_present: false,
+                            code: rumqttc::ConnectReturnCode::Success,
+                        },
+                    )))
+                })
             });
 
         client
@@ -900,14 +915,14 @@ mod tests {
                         predicate::eq("realm/device_id/control/consumer/properties".to_string()),
                         predicate::always(),
                     )
-                    .returning(|_topic, _qos| notify_success());
+                    .returning(|_topic, _qos| notify_success(SubAck::new(0,Vec::new())));
 
                 client
                     .expect_subscribe()
                     .once()
                     .in_sequence(&mut seq)
                     .with(predicate::eq("realm/device_id/org.astarte-platform.rust.examples.individual-datastream.ServerDatastream/#".to_string()), predicate::always())
-                    .returning(|_: String, _| notify_success());
+                    .returning(|_: String, _| notify_success(SubAck::new(0, Vec::new())));
 
                 // Client id
                 client
@@ -920,7 +935,7 @@ mod tests {
                         predicate::always(),
                         predicate::always(),
                     )
-                    .returning(|_, _, _, _| notify_success());
+                    .returning(|_, _, _, _| notify_success(AckOfPub::None));
 
                 // empty cache
                 client
@@ -933,7 +948,7 @@ mod tests {
                         predicate::always(),
                         predicate::eq("1"),
                     )
-                    .returning(|_, _, _, _| notify_success());
+                    .returning(|_, _, _, _| notify_success(AckOfPub::None));
 
 
                 // purge device properties
@@ -945,7 +960,7 @@ mod tests {
                         topic == "realm/device_id/control/producer/properties"
                             && *qos == QoS::ExactlyOnce
                     })
-                    .returning(|_, _, _, _| notify_success());
+                    .returning(|_, _, _, _| notify_success(AckOfPub::None));
 
                 // device property publish
                 client
@@ -953,9 +968,22 @@ mod tests {
                     .once()
                     .in_sequence(&mut seq)
                     .with(predicate::eq("realm/device_id/org.astarte-platform.rust.examples.individual-properties.DeviceProperties/sensor1/name".to_string()), predicate::always(), predicate::always(), predicate::always())
-                    .returning(|_, _, _, _| notify_success());
+                    .returning(|_, _, _, _| notify_success(AckOfPub::None));
 
                 client
+            });
+
+        // Catch other call to poll
+        eventl
+            .expect_poll()
+            .once()
+            .in_sequence(&mut seq)
+            .returning(|| {
+                Box::pin(async {
+                    tokio::time::sleep(Duration::from_millis(100)).await;
+
+                    Ok(Event::Incoming(rumqttc::Packet::PingReq))
+                })
             });
 
         let interfaces = [
@@ -994,7 +1022,7 @@ mod tests {
                     true,
                 )
                 .await
-                .unwrap(),
+                .expect("failed to configure transport provider"),
                 ClientId {
                     realm: "realm",
                     device_id: "device_id",