+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.10.0] - Unreleased
 
 ### Added
 
@@ -15,6 +15,38 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 
 - Bump MSRV to 1.78.0 [#395](https://github.com/astarte-platform/astarte-device-sdk-rust/pull/395).
+
+## [0.9.6] - 2025-03-06
+
+### Changed
+
+- Use a default CryptoProvider when defining a Tls configuration
+  [#425](https://github.com/astarte-platform/astarte-device-sdk-rust/pull/425).
+
+## [0.9.5] - 2025-03-04
+
+### Changed
+
+- Update `rumqttc` and `rustls` dependencies [#423]
+- Drop `openssl` as a dependency and deprecate the feature [#423]
+
+### Fixed
+
+- Fix a bug in the MQTT connection that will wait for the keep-alive timeout before connecting
+  [#403]
+
+[#423]: https://github.com/astarte-platform/astarte-device-sdk-rust/pull/423
+
+## [0.9.4] - 2025-02-27
+
+## [0.8.5] - 2025-02-27
+
+## [0.7.5] - 2025-02-27
+
+### Fixed
+
+- On reconnection send only device properties that are currently in the introspection
+  [#413](https://github.com/astarte-platform/astarte-device-sdk-rust/pull/413/)
 
 ## [0.9.3] - 2025-01-24